pyproject.toml
[tool.poetry]
name = "poltergeist"
version = "0.1.0"
description = ""
authors = ["Chris <qkymzh123@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
black = "^23.7.0"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
